+++ L91SuluFormBundle.php
@@ -2,6 +2,7 @@
 
 namespace L91\Sulu\Bundle\FormBundle;
 
+use L91\Sulu\Bundle\FormBundle\DependencyInjection\CompilerPass\DynamicListBuilderCompilerPass;
 use L91\Sulu\Bundle\FormBundle\DependencyInjection\CompilerPass\ListProviderCompilerPass;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
@@ -13,5 +14,6 @@ class L91SuluFormBundle extends Bundle
         parent::build($container);
 
         $container->addCompilerPass(new ListProviderCompilerPass());
+        $container->addCompilerPass(new DynamicListBuilderCompilerPass());
     }
 }