+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## dev-develop
+ 
+ - FEATURE     #121   Added dynamic form list and export
  - BUGFIX      #120   Fixed permission problem on false locale
  - BUGFIX      #117   Fixed limitation of content type select
  - FEATURE     #115   Added options for attachment validation